+++ src/System/Router/renderServerApp.tsx
@@ -1,6 +1,6 @@
 import { ArtsyRequest, ArtsyResponse } from "Server/middleware/artsyExpress"
 import { getServerParam } from "Utils/getServerParam"
-import { renderToString } from "react-dom/server"
+import { renderToStaticMarkup, renderToString } from "react-dom/server"
 import loadAssetManifest from "Server/manifest"
 import path from "path"
 import { ENABLE_SSR_STREAMING } from "Server/config"
@@ -8,6 +8,7 @@ import { getENV } from "Utils/getENV"
 import { ServerAppResults } from "System/Router/serverRouter"
 import { getWebpackEarlyHints } from "Server/getWebpackEarlyHints"
 import { RenderToStreamResult } from "System/Router/Utils/renderToStream"
+import { HtmlTemplate } from "HtmlTemplate"
 
 // TODO: Use the same variables as the asset middleware. Both config and sharify
 // have a default CDN_URL while this does not.
@@ -85,6 +86,12 @@ export const renderServerApp = ({
   }
 
   const statusCode = getENV("statusCode") ?? code
+
+  const htmlShell = renderToStaticMarkup(<HtmlTemplate {...options} />)
+
+  res.status(statusCode).send(htmlShell)
+
+  return
 
   res
     .status(statusCode)
